requirements.txt
Package            Version
------------------ ---------
certifi            2024.8.30
charset-normalizer 3.4.0
filelock           3.16.1
fsspec             2024.10.0
huggingface-hub    0.26.2
idna               3.10
Jinja2             3.1.4
MarkupSafe         3.0.2
mpmath             1.3.0
networkx           3.4.2
numpy              2.1.3
opencv-python      4.10.0.84
packaging          24.2
pdf2image          1.17.0
pillow             11.0.0
pip                24.3.1
PyPDF2             3.0.1
pytesseract        0.3.13
PyYAML             6.0.2
regex              2024.11.6
requests           2.32.3
safetensors        0.4.5
setuptools         75.3.0
sympy              1.13.1
tokenizers         0.20.3
torch              2.5.1
torchaudio         2.5.1
torchvision        0.20.1
tqdm               4.67.0
transformers       4.46.2
typing_extensions  4.12.2
urllib3            2.2.3
